# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs Untuk E-Learning Kampus",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reports/students/{id}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get student assignment and grade statistics",
                "operationId": "0a895835074fe86236f75d15939ee310",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Course"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/submissions/{id}/grade": {
            "put": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Update the grade of a submission",
                "operationId": "ae042d453980c272c447f8f81376f385",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "grade": {
                                        "type": "integer",
                                        "example": 85
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Grade updated and email sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubmissionResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation Error."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Submission not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Submission not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/submissions": {
            "post": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Upload a new submission",
                "operationId": "269db56b38031e1736e5c9162b88cdaa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "assignment_id": {
                                        "type": "integer"
                                    },
                                    "student_id": {
                                        "type": "integer"
                                    },
                                    "file_path": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "File uploaded and email sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubmissionResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation Error."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "File upload failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "File upload failed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Course": {
                "title": "Course",
                "properties": {
                    "id": {
                        "description": "Course ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Course name",
                        "type": "string"
                    },
                    "students_count": {
                        "description": "Number of students enrolled",
                        "type": "integer"
                    },
                    "assignments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Assignment"
                        }
                    }
                },
                "type": "object"
            },
            "Assignment": {
                "title": "Assignment",
                "properties": {
                    "id": {
                        "description": "Assignment ID",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Assignment title",
                        "type": "string"
                    },
                    "graded_submissions_count": {
                        "description": "Number of graded submissions",
                        "type": "integer"
                    },
                    "ungraded_submissions_count": {
                        "description": "Number of ungraded submissions",
                        "type": "integer"
                    },
                    "submissions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Submission"
                        }
                    }
                },
                "type": "object"
            },
            "Submission": {
                "title": "Submission",
                "properties": {
                    "id": {
                        "description": "Submission ID",
                        "type": "integer"
                    },
                    "student_id": {
                        "description": "Student ID",
                        "type": "integer"
                    },
                    "graded_at": {
                        "description": "Date and time when the submission was graded",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "SubmissionResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "assignment_id": {
                        "type": "integer"
                    },
                    "student_id": {
                        "type": "integer"
                    },
                    "file_path": {
                        "type": "string"
                    },
                    "score": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "scantum",
                "scheme": "bearer"
            }
        }
    },
...
}